moved getKineticEnergy to header and format fix, Ref. #62
